backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
httpx==0.28.1
google-generativeai==0.8.3
pywebpush==2.0.1
py-vapid==1.9.2
PyJWT[crypto]==2.10.1
supabase==2.11.0
pytest==8.3.4
pytest-asyncio==0.25.0
